font size fix for webkit browsers
For some unknown reason external webpages shown in an iframe are moved up because of the dynamically calculated font size.
This seems to only affect iframes and only webkit based browsers (safari, epiphany,...).
The fix for this to add an extra class to the container for the iframe which resets the font size to zero.
Testing showed that is does not have an effect on other engines or post types.